fixed up all import errors in refactoring of user layer into separate modules
