+++ lib9ml/python/nineml/user_layer/projection/grids.py
@@ -13,7 +13,7 @@
 """
 
 import math
-from nineml.geometry import Geometry
+from ...abstraction_layer.connection_generator.geometry import Geometry
 import nineml.user_layer
